ftdetect: Use `setfiletype`
From `:h setfiletype`:

" Set the 'filetype' option to {filetype}, but only if
not done yet in a sequence of (nested) autocommands.
This is short for:
:if !did_filetype()
    :  setlocal filetype={filetype}
    :endif
This command is used in a filetype.vim file to avoid
setting the 'filetype' option twice, causing different
settings and syntax files to be loaded. "